+++ chuang-framework/src/main/java/cn/zhangchuangla/framework/config/DruidDataSourceConfig.java
@@ -10,6 +10,7 @@ import javax.sql.DataSource;
 
 /**
  * DruidDataSource 配置
+ * @author Chuang
  */
 @Configuration
 public class DruidDataSourceConfig {
