--- chuang-framework/src/main/java/cn/zhangchuangla/framework/web/service/impl/SysPasswordServiceImpl.java
@@ -1,55 +1,0 @@
-package cn.zhangchuangla.framework.web.service.impl;
-
-import cn.zhangchuangla.common.core.config.property.SecurityProperties;
-import cn.zhangchuangla.common.core.enums.ResponseCode;
-import cn.zhangchuangla.common.core.exception.AccountException;
-import cn.zhangchuangla.common.redis.constant.RedisConstants;
-import cn.zhangchuangla.common.redis.core.RedisCache;
-import cn.zhangchuangla.framework.web.service.SysPasswordService;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-
-import java.util.concurrent.TimeUnit;
-
-/**
- * @author Chuang
- * <p>
- * created on 2025/2/25 15:42
- */
-@Slf4j
-@Service
-@RequiredArgsConstructor
-public class SysPasswordServiceImpl implements SysPasswordService {
-
-    private final RedisCache redisCache;
-    private final SecurityProperties securityProperties;
-
-
-    /**
-     * 密码错误尝试次数校验,如果超过设置的错误次数则会禁止登录
-     *
-     * @param username 用户名
-     */
-    @Override
-    public void PasswordErrorCount(String username) {
-        String key = RedisConstants.PASSWORD_ERROR_COUNT + username;
-        long maxRetryCount = securityProperties.getPasswordConfig().getMaxRetryCount();
-        if (maxRetryCount == -1) {
-            return;
-        }
-        Integer retryCount = redisCache.getCacheObject(key);
-
-        if (retryCount == null) {
-            retryCount = 0;
-        }
-        Long keyExpire = redisCache.getKeyExpire(key);
-        if (retryCount >= maxRetryCount) {
-            throw new AccountException(ResponseCode.ACCOUNT_LOCKED, "账号已被锁定,请在" + keyExpire + "秒后重新尝试");
-        }
-
-        Integer lockTime = securityProperties.getPasswordConfig().getLockTime();
-        redisCache.setCacheObject(key, ++retryCount, lockTime, TimeUnit.MINUTES);
-    }
-
-}
