+++ chuang-system/chuang-system-core/src/main/java/cn/zhangchuangla/system/mapper/SysConfigMapper.java
@@ -15,7 +15,7 @@ public interface SysConfigMapper extends BaseMapper<SysConfig> {
      * 分页查询
      *
      * @param sysConfigPage 分页对象
-     * @param request       查询参数
+     * @param sysConfigQueryRequest       查询参数
      * @return 返回分页数据
      */
     Page<SysConfig> listSysConfig(Page<SysConfig> sysConfigPage, @Param("request") SysConfigQueryRequest sysConfigQueryRequest);
