+++ chuang-system/chuang-system-core/src/main/java/cn/zhangchuangla/system/model/request/dept/SysDeptAddRequest.java
@@ -16,7 +16,6 @@ import lombok.Data;
 @Schema(name = "部门添加请求类", description = "部门添加请求类")
 public class SysDeptAddRequest {
 
-
     /**
      * 部门名称
      */
