+++ chuang-system/chuang-system-core/src/main/java/cn/zhangchuangla/system/model/vo/menu/MetaVo.java
@@ -21,7 +21,6 @@ import java.util.Set;
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
 public class MetaVo {
 
-
     /**
      * 设置该路由在侧边栏和面包屑中展示的名字
      */
