+++ chuang-system/chuang-system-core/src/main/java/cn/zhangchuangla/system/service/SysDeptService.java
@@ -16,7 +16,6 @@ import java.util.List;
  */
 public interface SysDeptService extends IService<SysDept> {
 
-
     /**
      * 获取部门列表
      *
