+++ chuang-system/chuang-system-core/src/main/java/cn/zhangchuangla/system/service/SysRoleService.java
@@ -18,7 +18,6 @@ import java.util.Set;
  */
 public interface SysRoleService extends IService<SysRole> {
 
-
     /**
      * 角色列表
      *
